Regress factors out of exposures and outcome

This is an optional method that regresses factor summary statistics out of exposure and outcome summary statistics after the factor analysis stage and before polygenic modeling and Bayesian regression. This will likely yield a conservative result. By default, famr_susie does not run this. See gather_sumstats for more details.

Usage

annihilate_factors(sumstats, factor_ss)

Arguments

sumstats

A list object, usually read by gather_sumstats. The list has three matrices with SNPs as rows and exposure traits as columns. These are "betas" (effect size estimates of SNPs on exposures), "stderrs" (standard errors of the betas), and "Z" (the z-scores, i.e. betas/stderrs). It also has three corresponding vectors betas_y, stderrs_y, and zscores_y for the outcome trait ("y"). It also has a vector 'pos' with the SNP names.

factor_ss

Like sumstats, this is a list object with betas, stderrs, and Z matrices, except for the factors instead of the exposures. Usually generated by generate_factors. It must also have an integer n_factors giving the number of factors. It must have the same set of SNPs as sumstats.

Value

A modified version of the sumstats object. The summary statistics for the exposures and outcome will have had the factors projected (regressed) out of them. It adds the fields orig_betas_y and orig_zscores_y, which are the values those fields had before annihilation, and annih_y, which is the portion of the outcome that was projected out. Thus, orig_betas_y = betas_y + annih_y in the returned object.