Factor-Augmented Mendelian Randomization with SuSiE (FAMR-Susie)

This is the main user-facing function that runs all parts of the FAMR-Susie method, described in the manuscript "Factor-based methods for robust high-dimensional Mendelian randomization". FAMR-Susie is intended for use in "high-dimensional" MR, where there are many genetically-correlated exposures, and simultaneously learns which of these exposures have a causal effect on the "outcome" trait. FAMR-Susie uses factor analysis to infer and correct for unobserved confounders, and uses sparse polygenic modeling and Bayesian regression to alleviate other challenges such as other violations of MR assumptions and sparsity. For more details on the method, please see the manuscript.

Usage

famr_susie(
  in_dir,
  ld_dir,
  y_gwas_file = "NONE",
  fa_method = "gfa",
  num_samples = 10000,
  n_iter = 30,
  susieL = 30,
  prune_thresh = 0.5,
  fa_prune_thresh = 0.1,
  final_prune_thresh = 0.1,
  annihilate = FALSE,
  idcol = 1,
  betacol = 2,
  secol = 3,
  header = TRUE
)

Arguments

in_dir

A directory with input rds files containing summary statistics for the exposures, with one file per locus. Each file should be a list object with three matrices with SNPs as rows and exposure traits as columns. These are "betas" (effect size estimates of SNPs on exposures), "stderrs" (standard errors of the betas), and "Z" (the z-scores, i.e. betas/stderrs). Optionally can also include corresponding vectors betas_y, stderrs_y, and zscores_y for the outcome trait ("y"); alternatively, those can be read in separately via the read_y_gwas function (see below). Finally, it has a vector 'pos' with the SNP names. See gather_sumstats for more information.

ld_dir

A directory with input rds files containing LD matrices corresponding to the SNPs in the summary statistics files, with one file per locus. These should simply be numerical nSNP-by-nSNP matrices.

y_gwas_file

An optional file path to the file with the outcome summary statistics. Can either be a TSV file with columns for the SNP name, effect size estimates, and standard errors (see idcol, betacol, secol), or a VCF file in the MRC IEU GWAS format, described here: https://github.com/MRCIEU/gwas-vcf-specification See also read_vcf in utils.R.

fa_method

The factor analysis (FA) method to be used. This defaults to 'gfa' and generally should not be changed, as FAMR-Susie currently only recognizes 'gfa'. If the user wishes to modify the code to allow different FA methods, they can specify those different methods here. Currently, if the user puts in something other than 'gfa', no factors are inferred. This can potentially be useful if the user wants to assess FAMR-Susie both with and without factors.

num_samples

The number of samples that were used to generate the exposure summary statistics. If this varies between the exposures, it can be something like the average – it does not have to be exact.

n_iter

The number of Expectation Maximization iterations to fit the cTWAS-style algorithm to learn prior effect probabilities for SNPs and exposures/factors. Default: 30.

susieL

The maximum number of effects (credible sets) allowed in the susie_rss runs. Default: 30.

prune_thresh

The absolute value correlation threshold used in the LD clumping step performed by FAMR-Susie before any further analysis. This is done to improve computational speed and stabilize results. We generally recommend a value between 0.3 and 0.9; the default is 0.5.

fa_prune_thresh

A stricter clumping threshold to generate a set of SNPs for input to factor analysis with GFA. Default: 0.1. It is not generally recommended to change this value much.

final_prune_thresh

A stricted clumping threshold to generate a set of SNPs to model in the cTWAS-style framework / susie_rss runs. By default it is 0.1, the same as fa_prune_thresh.

annihilate

Boolean; whether to regress factor summary statistics out of exposure and outcome summary statistics after the FA stage and before polygenic modeling and Bayesian regression. Setting this to TRUE will likely yield a conservative result. Default: FALSE.

idcol

Column of y_gwas_file with the SNP IDs (names). Default is 1, i.e. the first column. Not used if a VCF is provided.

betacol

Column of y_gwas_file with the SNP effect sizes (betas). Default is 2, i.e. the second column. Not used if a VCF is provided.

secol

Column of y_gwas_file with the SNP effect standard errors. Default is 3, i.e. the third column. Not used if a VCF is provided.

header

Boolean indicating whether the y_gwas_file has a header line, e.g. with column names. Default=TRUE. Not used if a VCF is provided.

Value

A list with some or all of the following elements:

pips

A list of the learned PIPs for each exposure, factor, and SNP.

zscores

A list of the marginal association z-scores learned for each exposure, factor, and SNP with the outcome.

R

A matrix of the estimated correlations between each pair of variables (exposures, factors, and SNPs).

susieres

The full susie_rss results object. See the susieR documentation for more details.

priors

A list containing the learned prior effect probabilities and prior effect variances learned for each class of variables (i.e. SNPs and exposures/factors).

posterior_mean

A list containing the posterior means learned for each variable by susie_rss.

posterior_var

A list containing the posterior variances learned for each variable by susie_rss.

factor_corrs

A matrix of correlations between the learned factor summary statistics and those of the exposures. Deprecated in favor of the R matrix.

Details

The famr_susie method expects summary statistics and linkage disequilibrium (LD) matrices as input. The user is expected to have split their data into independent loci before running famr_susie. Thus famr_susie expects a directory (in_dir) of summary statistics and another (ld_dir) for LD matrices, with one file each per locus. The data for the outcome phenotype may be included in the summary statistics files or supplied separately via the y_gwas_file argument. For the formats of these files, see the documentation for gather_sumstats and read_y_gwas, respectively. The remaining arguments are explained below.