Generate factors for FAMR-Susie

Generate factors given the summary statistics read in by gather_sumstats. Since the factors are usually learned with a more strictly LD-clumped set of summary statistics, this function also imputes the SNP-factor effect size estimates for the rest of the SNPs.

Usage

generate_factors(
  fa_method,
  sumstats,
  N = 10000,
  given_factors = "NONE",
  full_ss = c()
)

Arguments

fa_method

A string providing the factor analysis method to use. Currently this only recognizes 'gfa', but this could be extended to recognize other options in the future. If anything other than 'gfa' is passed in, the method assumes no factors should be generated.

sumstats

Input summary statistics to use for factor analysis, usually from gather_sumstats. A list object with the following elements:

betas

A matrix of estimated effect sizes of SNPs on exposures, with rows corresponding to SNPs and columns corresponding to exposures.

stderrs

A matrix of standard errors for the betas.

Z

A matrix of z-scores for the exposures, equal to betas divided by stderrs.

pos

A vector of SNP names.

N

Sample size of the studies used to generate GWAS summary statistics.

given_factors

DEPRECATED. A file path to factor summary statistics that have been pre-generated by the user. Used for internal testing purposes.

full_ss

Similar to sumstats, but contains the full set of SNPs to be used in further analysis, i.e. if prune_thresh is greater than fa_prune_thresh in gather_sumstats.

Value

A list object containing inferred summary statistics for factors, with the following elements:

betas

A matrix of estimated effect sizes of SNPs on exposures, with rows corresponding to SNPs and columns corresponding to exposures.

stderrs

A matrix of standard errors for the betas.

Z

A matrix of z-scores for the exposures, equal to betas divided by stderrs.

pos

A vector of SNP names.

n_factors

Number of inferred factors.