Run final Susie-RSS regression and return results

Runs Susie-RSS given prior inclusion probabilities and prior variances for each class of variables, and returns a comprehensive results object. Intended to be used after the cTWAS prior-fitting EM algorithm has been run.

Usage

get_results(zscores, R, priors = c(), L = 10)

Arguments

zscores

A numeric vector of z-score association statistics for the variables on the trait of interest.

R

A numeric matrix or sparse Matrix object of correlations between variables

priors

A list object with the following elements:

pi

A list containing the updated cTWAS prior inclusion probability (prior probability of effect on the trait) for each class of variables

sigma2

A list containing the updated cTWAS prior effect variances, conditioned on the variable having an effect, for each class of variables

n_vars

A list object giving the number of variables in each class, e.g. ('SNPs' = 10000, 'exposures' = 30)

L

The maximum number of effects (credible sets) Susie-RSS is allowed to fit.

Value

A list object with the following elements:

pips

A list object with elements for each class of variables, containing a numeric vector of PIPs for the variables of that class

posterior_var

A list object with elements for each class of variables, containing a numeric vector of posterior second moments (conditioned on the variable having an effect) for the variables of that class

posterior_mean

A list object with elements for each class of variables, containing a numeric vector of posterior mean effect size estimates for the variables of that class

susieres

A Susie-RSS results object returned by run_susie_with_priors; see the susieR documentation for more details.

zscores

Equal to the zscores parameter given as input.

R

Equal to the R parameter given as input.

L

Equal to the L parameter given as input.

sorted_pips

DEPRECATED. The unlisted (vector-format) pips sorted in decreasing order.