Prune GFA factors associated with only one exposure

GFA occasionally infers factors that are very strongly associated with only one exposure and not with others, so that they are essentially a proxy for that exposure. This helper function removes those factors heuristically, by removing factors that do not have an R^2 above some value ("thresh") with at least "min_count" exposures.

Usage

gfa_factor_prune_full(x_betas, gfares, thresh = 0.1, min_count = 2)

Arguments

x_betas

Numeric matrix of marginal effect size estimates ("betas") of SNPs on each exposure, with rows corresponding to SNPs and columns corresponding to exposures.

gfares

GFA results object, usually generated by run_gfa_full.

thresh

The R^2 threshold with exposures that the factors in gfares must meet to avoid being removed. Default: 0.1.

min_count

The number of exposures a factor must be sufficiently correlated with to avoid being removed. Default: 2.

Value

A GFA object (see GFA for more details); a modified version of gfares that contains only the factors that passed the filter.