Form correlation matrix of all variables

Generates a correlation matrix between all pairs of variables, including SNPs, exposures, and factors, given the weights learned by learn_weights in f amr.R and the LD between the SNPs. Can optionally take in wRw and Rgx (see below) if those have already been computed.

Usage

impute_exposure_corrs(ld, weights, wRw = c(), Rgx = c())

Arguments

ld

A numeric matrix or sparse Matrix object with the linkage disequilibrium (LD) values between each pair of SNPs, for the SNPs in the weights parameter/

weights

A numeric matrix of SNP weights on the exposures and factors, with rows corresponding to SNPs and columns to expsosures and factors.

wRw

A numeric matrix of associations between traits (exposures and factors), used for computing both Z and R. See precompute_zxy in famr.R for more details.

Rgx

A numeric matrix of inferred correlations between SNPs and traits (exposures and outcomes), with one row per SNP and one column per exposure/factor.See precompute_zxy in famr.R for more details.

Value

A numeric matrix or sparse Matrix object containing the inferred correlations between all SNPs, exposures, and factors.