Learn weights of SNP effects on exposures and factors

Learns the "weights" (joint effect sizes) of SNPs on each exposure and factor, using Susie-RSS. Uses the lambda parameters learned by learn_lambda for regularization. Adjusts other Susie parameters adaptively according to lambda, i.e. the "L" (number of effects) parameter and clumping threshold.

Usage

learn_weights(
  sumstats,
  ld,
  lambda,
  nome = F,
  N = 10000,
  L = 10,
  prune_min = 0.1,
  prune_max = 0.5
)

Arguments

sumstats

A list object, usually read by gather_sumstats. The list has three matrices with SNPs as rows and exposure traits as columns. These are "betas" (effect size estimates of SNPs on exposures), "stderrs" (standard errors of the betas), and "Z" (the z-scores, i.e. betas/stderrs). See gather_sumstats for more details.

ld

numeric LD matrix corresponding to the SNPs in sumstats

lambda

numeric vector of lambda parameters, one per trait, learned by learne_lambda.

nome

DEPRECATED. used only for internal testing. in this case, it automatically returns the "betas" as the weights, as it assumes a no measurement error (nome) model.

N

number of samples used to generate the exposure data, passed as an argument to susie-rss.

L

maximum number of effects (per-trait) susie-rss can model

prune_min

this sets the minimum LD clumping threshold allowed to be applied for any trait, after adjusting for the lambda parameter.

prune_max

this sets the maximum LD clumping threshold allowed to be applied for any trait, after adjusting for the lambda parameter.

Value

a numeric matrix of weights learned for each SNP on each exposure and factor, with one row per SNP and one column per exposure and factor.