Learn weights for SNPs, compute values for Z and R, merge data across loci

Essentially this is a large wrapper function that takes the summary statistics for exposures and factors and prepares the data necessary to impute Z-scores and the R matrix used in the cTWAS/Susie regression step. Merges sumstats for exposures and factors, learns weights on both exposures and factors, precomputes values needed to impute entries of the Z-scores vector and R matrix corresponding to exposures and factors, and LD clumps and merges all of this data into a single sumstats object and LD object.

Usage

learn_wts_precomp_merge(
  all_sumstats,
  factor_ss,
  N = 10000,
  nome = F,
  prune_thresh = 1
)

Arguments

all_sumstats

A list object, usually generated by gather_sumstats, with the following elements:

betas

A matrix of estimated effect sizes of SNPs on exposures, with rows corresponding to SNPs and columns corresponding to exposures.

weights

Currently this is the same as betas; later used to store learned weights of SNPs on exposures and factors.

stderrs

A matrix of standard errors for the betas.

Z

A matrix of z-scores for the exposures, equal to betas divided by stderrs.

betas_y

A vector of estimated effect sizes of SNPs on the outcome.

stderrs_y

A vector of standard errors for betas_y.

zscores_y

A vector of z-scores of SNPs on the outcome, equal to betas_y divided by stderrs_y.

pos

A vector of SNP names.

locus_idx

A vector matching SNPs to which locus they come from, indexed in the order the loci are read from in_dir.

lambda

A matrix of lambda values learned by learn_lambda for each exposure in each locus, where each row of the matrix corresponds to a locus and the columns are the exposures.

factor_ss

A list object containing inferred summary statistics for factors, usually from generate_factors, with the following elements:

betas

A matrix of estimated effect sizes of SNPs on factors, with rows corresponding to SNPs and columns corresponding to factors.

weights

Currently this is the same as betas; later used to store learned weights of SNPs on exposures and factors.

stderrs

A matrix of standard errors for the betas.

Z

A matrix of z-scores for the exposures, equal to betas divided by stderrs.

pos

A vector of SNP names.

n_factors

Number of inferred factors.

N

Number of samples that summary statistics were generated using.

nome

DEPRECATED. Boolean value indicating whether to assume a "no measurement error" model. Default: FALSE. Used for internal testing purposes.

prune_thresh

Numeric value between 0 and 1 indicating a final LD clumping threshold to be applied to the data before merging the loci. SNPs removed in this step will still be used to learn weights and compute Z and R, but will not be included in the final cTWAS/Susie regression.

Value

A list object with the following elements:

sumstats

A list object containing summary statistics for all exposures and factors; see the all_sumstats parameter for format. The "weights" element now contains weights learned by Susie-RSS.

ld

A sparse Matrix object with the LD for the SNPs in the sumstats object.

numerator

A numeric vector representing the numerator of the Z-scores for each exposure and factor.

wRw

A numeric matrix of associations between traits (exposures and factors), used for computing both Z and R

Rgx

A numeric matrix of inferred correlations between SNPs and traits (exposures and outcomes), with one row per SNP and one column per exposure/factor.