Merge outcome and exposure summary statistics

Given exposure summary statistics and outcome summary statistics, intersects the SNPs, then merges them into a single object. Used if read_y_gwas is run, i.e. if outcome summary statistics are not already included in the exposure summary statistics files.

Usage

merge_outcome(sumstats, y_gwas, ld = c())

Arguments

sumstats

A list object with the following elements, i.e. summary statistics for exposures:

betas

A matrix of estimated effect sizes of SNPs on exposures, with rows corresponding to SNPs and columns corresponding to exposures.

weights

Currently this is the same as betas; later used to store learned weights of SNPs on exposures and factors.

stderrs

A matrix of standard errors for the betas.

Z

A matrix of z-scores for the exposures, equal to betas divided by stderrs.

pos

A vector of SNP names.

y_gwas

A list object with the following elements, i.e. summary statistics for the outcome:

names_y

A vector of SNP IDs (names)

betas_y

A vector of estimated effect sizes of SNPs on the outcome.

stderrs_y

A vector of standard errors for betas_y.

zscores_y

A vector of z-scores of SNPs on the outcome, equal to betas_y divided by stderrs_y.

ld

A numeric LD matrix between the SNPs in sumstats.

Value

A list object with the following elements:

sumstats

A list with the fields of both sumstats and y_gwas (except names_y, which is merged into pos), containing the merged summary statistics

ld

A numeric LD matrix for the SNPs in the new sumstats object; the same as the input LD matrix except without the SNPs that are unmeasured in y_gwas.