Merge sumstats for a locus into overall sumstats

This function is meant to be called from several functions in famr.R that loop through loci. Given an object containing summary statistics merged from all loci so far (all_sumstats) and summary statistics from a new locus, merges these two and returns a new summary statistics object. Also used to merge exposure and factor summary statistics.

Usage

merge_sumstats(all_sumstats, locus_sumstats, by_row = T)

Arguments

all_sumstats

A summary statistics list object, containing some or all of the following fields (only betas, stderrs, weights, Z required):

betas

A matrix of estimated effect sizes of SNPs on exposures, with rows corresponding to SNPs and columns corresponding to exposures.

stderrs

A matrix of standard errors for the betas.

Z

A matrix of z-scores for the exposures, equal to betas divided by stderrs.

betas_y

A vector of estimated effect sizes of SNPs on the outcome.

stderrs_y

A vector of standard errors for betas_y.

zscores_y

A vector of z-scores of SNPs on the outcome, equal to betas_y divided by stderrs_y.

pos

A vector of SNP names.

locus_idx

A vector matching SNPs to which locus they come from, indexed in the order the loci are read from in_dir.

lambda

A matrix of lambda values learned by learn_lambda in famr.R for each exposure in each locus, where each row of the matrix corresponds to a locus and the columns are the exposures.

locus_sumstats

Like all_sumstats, but for the current locus only.

by_row

Boolean indicating whether the sumstats should be merged by row or by column. The latter is used when merging exposure and factor summary statistics rather than merging different loci.

Value

A list object in the same format as all_sumstats, with the merged summary statistics from all_sumstats and locus_sumstats.