Run Susie-RSS with cTWAS-inferred priors

A wrapper function to run Susie-RSS with priors for each variable given by the cTWAS algorithm at its current EM iteration.

Usage

run_susie_with_priors(zscores, R, prior_pi = c(), prior_sigma2 = c(), L = 10)

Arguments

zscores

A numeric vector of z-score association statistics for the variables on the trait of interest.

R

A numeric matrix or sparse Matrix object of correlations between variables

prior_pi

A list containing the updated cTWAS prior inclusion probability (prior probability of effect on the trait) for each class of variables (SNPs or exposures/factors)

prior_sigma2

A list containing the updated cTWAS prior effect variances, conditioned on the variable having an effect, for each class of v ariables (SNPs or exposures/factors)

L

The maximum number of effects (credible sets) Susie-RSS is allowed to fit.

Value

A Susie-RSS results object. See the susieR documentation for more details.