Subset summary statistics objects

Subset a summary statistics object ("ss") according to the indices provided by idx, i.e. retain only the SNPs at the positions given by idx.

Usage

subset_sumstats(ss, idx)

Arguments

ss

A summary statistics list object, containing some or all of the following fields (only betas, stderrs, weights, Z required):

betas

A matrix of estimated effect sizes of SNPs on exposures, with rows corresponding to SNPs and columns corresponding to exposures.

stderrs

A matrix of standard errors for the betas.

Z

A matrix of z-scores for the exposures, equal to betas divided by stderrs.

betas_y

A vector of estimated effect sizes of SNPs on the outcome.

stderrs_y

A vector of standard errors for betas_y.

zscores_y

A vector of z-scores of SNPs on the outcome, equal to betas_y divided by stderrs_y.

pos

A vector of SNP names.

locus_idx

A vector matching SNPs to which locus they come from, indexed in the order the loci are read from in_dir.

annih_y

A vector containing the portion of betas_y projected out by factors; see annihilate_factors in famr.R for more details

orig_betas_y

A vector containing the original betas_y if the latter was modified by annihilating factors; see annihilate_factors in famr.R for more details

orig_zscores_y

Similar to orig_betas_y but for zscores_y

idx

A vector of SNP indices to retain from ss

Value

A modified version of ss with only the SNPs specified by idx kept.