Update posterior parameters after running Susie-RSS

Given Susie-RSS results, updates the PIP (alpha) and posterior mean (post_mu) and posterior variance (tau2) of each variable in each "class" (SNPs and exposures/factors). Part of the cTWAS Expectation-Maximization (EM) algorithm to fit class-specific prior parameters for Susie-RSS.

Usage

update_alpha_tau2(res, n_vars)

Arguments

res

Susie-RSS results object generated by run_susie_with_priors; see the susieR documentation for more details. Most relevantly for this function, it contains the following elements:

pip

A numeric vector of Posterior Inclusion Probabilities (PIPs) for each variable

mu

A numeric matrix of posterior mean estimates for each variable in each credible set, with one row per credible set and one column per variable

mu2

Similar to mu, but for posterior second moments

alpha

A numeric matrix of PIPs for each credible set, with one row per credible set and one column per variable

n_vars

A list object giving the number of variables in each class, e.g. ('SNPs' = 10000, 'exposures' = 30)

Value

A list object with the following elements:

alpha

A list object with elements for each class of variables, containing a numeric vector of PIPs for the variables of that class

tau2

A list object with elements for each class of variables, containing a numeric vector of posterior second moments (conditioned on the variable having an effect) for the variables of that class

post_mu

A list object with elements for each class of variables, containing a numeric vector of posterior mean effect size estimates for the variables of that class