Update cTWAS priors

Given updated posterior parameters learned from running Susie-RSS given the previous cTWAS priors, update the cTWAS prior inclusion probability and prior variance for each class of variables (i.e. SNPs and exposures/factors). Part of the cTWAS EM algorithm.

Usage

update_pi_sigma2(alpha, tau2)

Arguments

alpha

A list object with elements for each class of variables, containing a numeric vector of PIPs for the variables of that class. Usually generated by update_alpha_tau2.

tau2

A list object with elements for each class of variables, containing a numeric vector of posterior second moments (conditioned on the variable having an effect) for the variables of that class. Usually generated by update_alpha_tau2.

Value

A list object with the following elements:

pi

A list containing the updated cTWAS prior inclusion probability (prior probability of effect on the trait) for each class of variables

sigma2

A list containing the updated cTWAS prior effect variances, conditioned on the variable having an effect, for each class of variables