Write FAMR-Susie results to a file

Writes FAMR-Susie results to a file, and optionally a simple results summary to an additional text file. Mainly intended to be used internally for simulations.

Usage

write_famr_res(out, famr_res, K, fa_method, for_sim = F)

Arguments

out

Base name for output files

famr_res

An FAMR-Susie results object; see the famr_susie method in famr.R for more details.

K

The number of exposures in the analysis.

fa_method

Which factor analysis method was used to infer factors for FAMR-Susie.

for_sim

A boolen indicating whether to generate an additional text file output for use with internal simulations.

Value

nothing